requirements.txt
Flask
Flask-APScheduler
azure-storage-queue
python-dotenv
pyodbc
azure-identity